Update patch set 1
Abandoned

The tests pass unaligned buffers to the _c used as a reference. That could be fixed, but this wouldn't work in high-bitdepth as those functions rely on the _c for situations that would cause overflow. It's going to be better to have full NEON implementations.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon