Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> (1 comment)
> 
> > Patch Set 9:
> > 
> > (1 comment)
> > 
> > does this report the error or fail in any configuration?
> 
> This will fail in asan.
> 

I'm not sure why the regular run isn't catching it, asan is used for the unit tests runs. Is it 32-bit only? In either case it should be brought in as DISABLED_ until the fix is in place

Patch-set: 10
Reviewer: James Zern <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>