Update patch set 5
Patch Set 5:

(3 comments)

> What would have to be done to update the test system for avx512? Are there any potential showstoppers (e.g. skylake-x not considered stable enough.)?

We run the test on virtualized machines (google cloud) so we don't really have the ability to swap out newer hardware. This isn't to say we can't add the functions, but testing will be tricky. With avx2 we ended up needing to fix some bugs when we tried to enable it.

Per James' comment, it looks like we could add emulator support. I'm a little concerned about speed, but if we restrict it to just avx512 tests it probably won't be an issue.

None of this blocks accepting this patch. I will need to update things like the chromium build to disable avx512 until we have good testing in place.

Patch-set: 5