Update patch set 18
Patch Set 18:

OK, I was assuming I could link both to one ABI bump if necessary.

I am still not clear though why we need compile time detection of this at all if it does not change the interface. If an older version of the libvpx library is called by an application that uses the film mode then libvpx will just not recognize the mode and I assume will report an error for an out of range input value. If a new version of libvpx is called by an application that does not know about the new allowed value to the enum then that will work fine because the other allowed values are the same as they were before.

I am still confused by the way we handle things especially the different way in which "extra" config is handled vs those parameters in the main config data structure. For example, the extra config parameters seem to be set up much later so they are not available as things stand when the second pass init function is called.

I can see that there is a problem with something that changes or adds in a new control such as the corpus vbr both for the case where an old libvpx library is called by an application that expects the control to be there and the case of a new library called by an application that does not know about the new control.

I had  assumed that applications using libvpx were supposed to poll the ABI version and only work if the return value is the one they expect, and that was the limit of the protection. Alternatively are applications supposed to support all earlier libvpx versions or are latter libvpx versions supposed to be backwards compatible. Should new controls or changes to controls inside libvpx be hidden under tests of the ABI version so that it is possible to build a compatible earlier version by resetting the abi version number to a lower value? As far as I can tell this is not the case.

Patch-set: 18
Tag: mailMessageId=<CALRks=Qu8mS4s1hgyM7YtUNyKsy5zGLkQGq6Bkf5+gJue6WqHg@mail.gmail.com>