Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This patch adds support for the analyzer backend. I'm not too familiar with libvpx, so I could use some help to clean it up, especially getting the includes right.
> 
> https://people.xiph.org/~mbebenita/analyzer_beta/index.html?d=https://people.xiph.org/~mbebenita/analyzer_beta/.media/vp9.js&f=https://people.xiph.org/~mbebenita/analyzer_beta/.media/vp9.ivf

I was going to try and take a look at this.  So first I thought to try it.  

I downloaded the patch and built via the shell script.   

Then I tried to load with the prebuilt AOMAnalyzer and got something like Decoder load failed.  

So I tried the github version ( in case the prebuilt didn't include something ).    

I ran that through electron in dev mode and grabbed the console output : 

and got something like : 

Failed to open /tmp/input.ivf for reading.

Along with a long list of of options and then some fails.   

I wanted to see it working before starting review.    

Please help?
Jim

Patch-set: 1
CC: James Bankoski <1000963@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>