Update patch set 1
Patch Set 1:

I forgot the --ivf which is necessary even if you do -o t.ivf   :)  Jim

ps. it looks good but a few things are off - eg in vp9 the transform type is tied to the mode so if you do v_pred you'll do an adst but this always shows dct_dct.    a few other things like that ...    but in general it works well  Thanks for doing this!

Patch-set: 1
Tag: mailMessageId=<CANzSYSdrzc=OrG+y0kxNiR-nua1DPPbZnUqDjLUDUdLbDT+mtA@mail.gmail.com>