Update patch set 1
Patch Set 1:

Well I am happy the try and make it less so.

I did try a weird animation that we had previously used in testing repeating. That did not trigger any false key frames but did pick up a couple we previously missed.

Our existing code was requiring 25% intra coded blocks in the first pass and the problem was that this was not happening in at least half the cases here. You could obviously extend the window to require static for x frames either side but how long does an image need to remain static in both directions for it to be a slide show.

Patch-set: 1
Tag: mailMessageId=<CALRks=R2=brj-cNcacit0e8WpBawG2x50uBvxmAn8RBVJUSxmA@mail.gmail.com>