Update patch set 1
Patch Set 1:

I am going to put up another patch which adds a second check as testing by yclin does appear to show cases with false positives even though there were none in our own test sets.

Having looked at the sample slide show he gave me it is also the case that at the transitions the intra and coded error are very similar. As such I have changed the slide transition test to require this to be the case.

return ((this_frame->intra_error / this_frame->coded_error) < 1.5) &&

(this_frame->coded_error > (last_frame->coded_error * 5.0)) &&

(this_frame->coded_error > (next_frame->coded_error * 5.0));

This additional clause may be a little too agressive and could mean that we still miss some transitions if the format of the images is the same and small, but this could still be tuned further later if we find cases that justify it and where the intra error is much higher than the inter error missing the key frame won't be such a disaster anyway.

Cases such as the frame upsampling I mentioned and Jim's giant text and anime will be unlikely to trigger unless it is almost as easy to intra code the frame, in which case a key frame may still be a good choice.

For example if the frame is only changing every 5 frames and the change is large, then coding a key frame every 5 and then almost nothing for the remainder may be a very good outcome.

Patch-set: 1
Tag: mailMessageId=<CALRks=TYans-c3WvwTQo27VezuzuAePcDdjbEhOGQ1dmokPsJQ@mail.gmail.com>