Update patch set 1
Patch Set 1:

was there much change in code generation? I know in the past on certain platforms people have recommended against the array setup as it inhibited the compiler. we use this pretty extensively, but I think this would be a good case to check modern clang and gcc output.

Patch-set: 1
CC: James Zern <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>