Update patch set 1
Patch Set 1:

Yes. Infact for larger formats the limit suggested by Aki of 250 bits per MB actually converges with the maximums allowed for those levels.

  rc->max_frame_bandwidth =
      VPXMAX(VPXMAX((cm->MBs * MAX_MB_RATE), MAXRATE_1080P), vbr_max_bits);

It is also worth noting here that if no level is specified but a very high rate is specified on the command line a higher value is allowed (because the user has explicitly asked for it).

Patch-set: 1
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>