Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I tried the poc on oc-mr1-dev and it could reproduce exactly the same output as in the bug. I wrapped the code where the bug happens in the motion vector limitation, and the crash seems to disappear (instead, the log shows "can't initialize encoder" instead of crashing).
> 
> Do we need to add a test to expose this first?

As noted offline, if the source is public we can add it to the test set. In the end the encoder should produce output for any yuv input, so there's a bit more to this bug I'd think.

Patch-set: 1
Reviewer: James Zern <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>