Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hey Bruce, given your understanding of floating point and visual studio, do you know if 'emms' is required to clean mmx state on 64 bit builds?

Ah, MMX. Such a delicious hack, layered on top of the wonderfully hackiness of x87.

emms/_m_empty/_mm_empty is needed anytime you switch from using MMX to x87. The x64 model was intended to be a clean break from the horrors of x87 so there should be no x87 instructions generated in x64 code - x64 compilers will not generate them. Therefore it should be fine to skip _m_empty for x64.

Corroborating evidence for this would be that _m_empty isn't documented for x64 (as mentioned in the CR comments) and not available for x64 (compile errors).

It is *possible* to use x87 instructions in an x64 process but you'd have to use assembly language to do it and nobody does. If somebody did then I think they would be responsible for starting out their x87 code with an emms instruction, since they are the anomaly.

Omitting it in x64 lgtm.

Patch-set: 2
Reviewer: Bruce Dawson <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>