Delete invalid assert.
Delete assert that is not valid in all cases.

This can occur if the last group in a clip is a GF only
group. Here the frame count reflects the nominal
positioning of the "next" GF (were it to exist) one
frame beyond the of the end of the clip.

Change-Id: I0d36b83de0ab478dab032599ee7df7fff4a35cd5