vp9: Initialize source variance in nonrd-pickmode.
It is already initialized at superblock level, but since
it is computed per coding block, based on some speed features,
better to initialize it in pick_inter.

No change in behavior, as currently the speed features
that enable use of source_variance in pick_inter are fixed
at the frame-level.

Change-Id: Ic787ac2f389ba1bced98716096e7b5cffba856a7