test/svc_end_to_end_test: fix SetConfig() signature
make the parameter constant to match the base class and mark the
function virtual. virtual is used to match the rest of the code base,
but now that c++11 is required all such functions could be changed to
override.

since:
bb3a82ec3 vp9 svc: add test for scaling partition on 1080p crash.

Change-Id: I4717f0116a231ea954b34da9cfec69c462c21699