Update patch set 1
Patch Set 1:

I don't think there is necessarily a mystery here as multi-threaded version often has a higher rate. Hence it is not necessarily the case that it is better, just different. If PSNR was higher and rate lower then that might indicate it was better but PSNR a little better and rate a little higher on a single data point does not.

Even so, we know that sometimes we get oddities because a small change causes the codec, for example, to cross over a boundary/threshold in terms of the choice of the initial active Q range (min-max)  or the length of a particular ARF group, and this causes the clip to be unexpectedly better or worse.

Patch-set: 1
Tag: mailMessageId=<CALRks=SMA_VKbLT_5uQuKR81Fw5AHDLikKC_KndUyJh8VWPjTg@mail.gmail.com>