Update patch set 1
Patch Set 1:

I tested this quite a lot and did not see much material difference with or without on any of the clips I tested (as per my comments on the original patch). However, I could see that in some circumstances it might cause a false positive or false negative where the previous head did not. That said I suspect that you could find clips where the converse is true and this patch might give a gain.

For a short animated clips such as the one in question a couple of wrong decisions on key frames can easily cause a big regression.

I am fine with taking this out but are we sure that against the couple where there was regression there are not others that gain ?

Patch-set: 1
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>