Refactor Neon highbd_convolve8 kernels
Move narrowing shift and max value clipping into the 4-pixel-output
kernel. As well as cleaning up the code quite a bit, this also
improves performance by 5-10% as it eliminates the implied top /
bottom register shuffling of the previous approach.

Also clean up the formatting and magic numbers in the 8-pixel-output
kernel.

Change-Id: I77a5e9e317ef4097f187330d4b32973022ba573f