Add vp8_ prefix to sem_* macro names
The sem_* macros do not behave exactly like the POSIX sem_* functions.
Add the vp8_ prefix to the sem_* macro names to make it clear that they
are not the POSIX sem_* functions. Another reason for adding the vp8_
prefix is that we need to wrap sem_wait() (to handle EINTR) on the Unix
platforms that have real sem_wait() function.

Handle EINTR in the Unix (non-Apple) definition of vp8_sem_wait().

Change-Id: I3df02a30f851d41691a55cf7a84aa2ff054bba9c