rtc-vp9: Disable svc_use_low_part for screen
Possible fix for issue below. The speed feature disabled
is only used for 3 spatial layers and at least 2 temporal.
The impact on speed is expected to be small, ~2%, so ok
to disable for now and see if it clears the issue.

Bug: 366146260

Change-Id: I94ab991d583cc2ce758db337abbbb463a65f0767
(cherry picked from commit 696c488d3581ce5258d4e01093f5f50840e2e40c)