rtc-vp9: Always disable svc_use_low_part
Possible fix for issue below. It was only disabled
for screen in a previous change, but we force it off
always to check if it clears the issue.

The speed feature disabled is only used for 3 spatial
layers and at least 2 temporal. The impact on speed is
expected to be small, ~2%, so ok to disable for now and
see if it clears the issue.

Bug: 366146260
Change-Id: If7af006425e1e0ef297b9d6466507ea4c90ddb6f