vpx_image: add allocation guidance for vpx_img_wrap
The image data buffer provided to `vpx_img_wrap()` must have a minimum
allocation size that satisfies the requirements of the `fmt`, `d_w`,
`d_h`` and `stride_align` parameters. This change adds a description of
how to calculate the size based on the one done in `img_alloc_helper()`.

Bug: 442142951
Change-Id: I06b40247e199f6683209fb2b6ade9616c5b63da7